WebGLRenderer: Unify cubemaps and cubeuvmaps into WebGLEnvironments.
